impl: more dbi

+++ src/main/kotlin/app/index_it/data/sources/db/dbi/user/EmailVerificationDBI.kt
@@ -6,7 +6,7 @@ import app.index_it.data.models.user.UserDto
 import app.index_it.data.sources.db.dbi.DBI
 
 interface EmailVerificationDBI : DBI {
-    suspend fun count(id: IxId<UserDto>): Int
+    suspend fun count(id: IxId<UserDto>): Long
     suspend fun save(emailVerificationDto: EmailVerificationDto)
     suspend fun get(token: String): EmailVerificationDto?
     suspend fun deleteAll(id: IxId<UserDto>)
